app-no-pillow.py
# app.py - Pillow-free version for deployment
# This version removes PIL dependency and handles images through Google AI directly

import os
import json
import sqlite3
import time
import hashlib
import secrets
from datetime import datetime
from functools import wraps
from flask import Flask, request, jsonify, render_template, render_template_string, redirect, url_for, flash, session
# from PIL import Image, ExifTags  # Commented out for deployment
from dotenv import load_dotenv
import google.generativeai as genai

# Compatibility shim for Flask/Werkzeug versions
try:
    from werkzeug.wrappers import Response as _WerkzeugResponse
    _orig_set_cookie = _WerkzeugResponse.set_cookie
    def _set_cookie_compat(self, *args, **kwargs):
        kwargs.pop('partitioned', None)
        return _orig_set_cookie(self, *args, **kwargs)
    _WerkzeugResponse.set_cookie = _set_cookie_compat
    if hasattr(_WerkzeugResponse, 'delete_cookie'):
        _orig_delete_cookie = _WerkzeugResponse.delete_cookie
        def _delete_cookie_compat(self, *args, **kwargs):
            kwargs.pop('partitioned', None)
            return _orig_delete_cookie(self, *args, **kwargs)
        _WerkzeugResponse.delete_cookie = _delete_cookie_compat
    try:
        from flask import wrappers as _flask_wrappers
        if hasattr(_flask_wrappers, 'Response'):
            _flask_wrappers.Response.set_cookie = _set_cookie_compat
    except Exception:
        pass
except Exception:
    pass

# Continue with rest of app.py but with GPS extraction disabled
# Since Pillow is removed, GPS from images will be skipped
# Device GPS will still work from JavaScript

load_dotenv()
app = Flask(__name__)
app.config['JSON_SORT_KEYS'] = False
app.secret_key = os.getenv('SECRET_KEY', secrets.token_hex(32))

# [Rest of your app.py content with GPS extraction function modified]

def get_gps_coordinates(image_stream):
    """Simplified GPS extraction - returns None without Pillow"""
    return {"latitude": None, "longitude": None}

# NOTE: This is just the header - you'd need to copy the rest of your app.py
# This shows how to handle the Pillow dependency issue