requirements-binary.txt
# requirements.txt with binary-only installation
--only-binary=Pillow
Flask==2.3.3
Pillow
python-dotenv==1.0.0
google-generativeai==0.3.2
Werkzeug==2.3.7
gunicorn==21.2.0